LayoutTests/fast/exclusions/shape-inside/shape-inside-polygon-padding-002-expected.txt

The blue diamond outline should be filled by cross made of 5 green squares.

X XXX X
PASS shapeInsideRect('a').top is 150
PASS shapeInsideRect('a').width is 100
PASS shapeInsideRect('b').top is 250
PASS shapeInsideRect('b').width is 300
PASS shapeInsideRect('c').top is 350
PASS shapeInsideRect('c').width is 100
